@mkhorton 2020-04-03
This is great, thanks @mturiansky! Optimizations always very welcome :-)

Now there's the question of why large arrays of floats are stored in ascii to start with ... perhaps time to kindly ask the VASP devs for a native hdf5 output.