@mturiansky 2020-04-03
> Now there's the question of why large arrays of floats are stored in ascii to start with ...

but ascii has _\~compatibility\~_...

/s