@mkhorton 2020-04-03
And also the question of why the ELFCAR and AECCAR are apparently printed in a lower precision. Perhaps there's a good technical reason for this, I'm not sure.